Add XML doc
