--- com.unity.netcode.gameobjects/Editor/InScenePlacedPrefab.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: e9b29e08242545899dfb65e4fb568beb
-timeCreated: 1779219564
