+++ com.unity.netcode.gameobjects/Editor/InScenePlacedProcessor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cda0fb70bdcd545a7983ca78f516bcff
